Fixed edgecase relating to transferring from certain addresses
Source addresses where the lowest word is 0 were affected. Subtracting
the low word from 0 to get the remaining words in the 128KB block would
fail if the low word is 0, since it would give a result of 0. In reality
it would mean there were 0x10000 words remaining.

This edgecase causes a nasty glitch in the current skdisasm if built
with Sonic3_Complete set, and ArtUnc_Obj_Lightning_Shield_Sparks is
aligned to 0x19FFFE: travelling partway through AIZ Act 1 causes the
entire VRAM to corrupt.